xcb: delay showing tray icon window until it is embedded
Origin: upstream, https://code.qt.io/cgit/qt/qtbase.git/commit/?id=612953a626ec21b8
Bug: https://bugs.debian.org/775398
Bug: https://bugs.debian.org/847665
Last-Update: 2016-12-26

Otherwise there is a race condition: when the tray implementation
gets around to embedding the window, if it was already shown,
it will be unmapped, embedded, and then remapped. Some tray
implementations will resize the tray icon to 1 pixel wide in
that case. We also never want to show a window that was intended
for the tray in any other location, so it's better that it remain
invisible until we are sure it is embedded.

Gbp-Pq: Name xcb_delay_showing_tray_icon_window_until_it_is_embedded.patch