fix crash when text contains too many directional chars
Origin: upstream, a concatenation of two commits:
 https://code.qt.io/cgit/qt/qtbase.git/commit/?id=79f2a9e666a241c5
 https://code.qt.io/cgit/qt/qtbase.git/commit/?id=1232205e32464d90
Last-Update: 2019-10-29

In case a text to be layouted contains more than 128 directional characters
it causes the application to crash.

The function initScriptAnalysisAndIsolatePairs() collects information of
RTL/LTR chaaracters into vector "isolatePairs". The size of the vector is
capped to 128. Later the function generateDirectionalRuns() iterates
the text again and tries to access items from the previously capped vector
above the upper bound.

Gbp-Pq: Name cve-2019-18281.diff